Rewrote node_match more efficiently. Remove ! operation for or/and branches. Make conntrack_create_entry fail if no match_priv is created. Fix warning on x86_64. Fix match reference count when destroying a rule.

+++ src/common.c
@@ -105,8 +105,8 @@ int frame_alloc_aligned_buff(struct frame *f, int length) {
 
 	int total_len = length + ic.buff_align_offset + 4;
 	f->buff_base = malloc(total_len);
-	f->buff = (void*) (((int)f->buff_base & ~3) + 4 + ic.buff_align_offset);
-	f->bufflen = total_len - ((int)f->buff - (int)f->buff_base);
+	f->buff = (void*) (((long)f->buff_base & ~3) + 4 + ic.buff_align_offset);
+	f->bufflen = total_len - ((long)f->buff - (long)f->buff_base);
 
 	return POM_OK;
